\name{cMonkey.bigdata-package}
\alias{cMonkey.bigdata-package}
\alias{cMonkey.bigdata}
\docType{package}
\title{
Additional example (big)data for use with the cMonkey integrated biclustering algorithm
}
\description{
Additional example (big)data for use with the cMonkey integrated biclustering algorithm
}
\details{
\tabular{ll}{
Package: \tab cMonkey.bigdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-04\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
\url{}
}
\author{
David J Reiss, Institute for Systems Biology

Maintainer: \email{dreiss@systemsbiology.org}
}
\references{
 
}
\keyword{ package }
\seealso{
\url{}


}
\examples{
 
}
