\name{ecoli}
\alias{ecoli}
\docType{data}
\title{
##   ## data name/kind ... ##
}
\description{
##  ## A concise (1-5 lines) description of the dataset. ##
}
\usage{data(ecoli)}
\format{
  The format is:
List of 7
 $ organism         : chr "eco"
 $ k.clust          : num 300
 $ rsat.species     : chr "Escherichia_coli_K12"
 $ ratios.old       :'data.frame':	3859 obs. of  82 variables:
  ..$ cipro_0_1                   : num [1:3859] -2.06 -1.91 -1.57 -0.01 0.19 -0.41 -0.65 -1.43 -0.47 -0.08 ...
  ..$ crp_0_1                     : num [1:3859] -0.6 -1.05 -0.59 0.21 0.55 0.15 -0.16 -0.27 -0.12 0.14 ...
  ..$ csp_0_1                     : num [1:3859] 0.01 -0.31 0.16 0.53 0.56 0.1 0.2 0.18 0.2 0.46 ...
  ..$ dps_1_1                     : num [1:3859] -3.27 -2.17 -2.65 -0.62 0.51 ...
  ..$ dps_2_1                     : num [1:3859] -3.15 -1.87 -2.23 -0.37 0.74 -1.75 -0.2 0.12 0.62 0.36 ...
  ..$ dps_3_1                     : num [1:3859] -3.63 -2.43 -2.7 0.486 0.57 ...
  ..$ gcurve2_1_1                 : num [1:3859] -0.81 -0.65 -0.37 -0.47 0.96 -0.02 -0.16 -0.36 -0.42 1.26 ...
  ..$ gcurve2_2_1                 : num [1:3859] -3.25 -1.72 -2.46 -0.05 1.1 -1.62 -0.28 0.25 0.1 1.6 ...
  ..$ gcurve2_3_1                 : num [1:3859] -3.2 -1.71 -2.49 0.03 0.76 -2.02 -0.07 -1.01 0.73 0.84 ...
  ..$ gcurve2_4_1                 : num [1:3859] -3.2 -1.64 -2.65 -1.22 0.78 ...
  ..$ heat50_10_1                 : num [1:3859] -4.59 -3.02 -3.06 -1.04 -0.36 -0.59 -0.41 -0.02 -0.44 -1.17 ...
  ..$ hns_0_1                     : num [1:3859] -1.17 -1.16 -0.86 -0.13 1.16 -0.68 -0.16 -0.08 0.25 0.83 ...
  ..$ hup_0_1                     : num [1:3859] -0.73 -1.2 -0.57 0.43 0.17 -0.14 -0.04 -0.38 0.47 0.22 ...
  ..$ pH2_0_1                     : num [1:3859] -1.05 -0.81 -0.87 0.37 0.46 -0.24 -0.05 -0.15 0.57 -0.15 ...
  ..$ mean.log2RatioNormalized.84 : num [1:3859] NA NA NA 1.374 0.131 ...
  ..$ mean.log2RatioNormalized.86 : num [1:3859] NA NA NA 0.268 -1.301 ...
  ..$ mean.log2RatioNormalized.87 : num [1:3859] NA NA NA 0.228 -1.211 ...
  ..$ mean.log2RatioNormalized.88 : num [1:3859] NA NA NA 0.533 -1.116 ...
  ..$ mean.log2RatioNormalized.90 : num [1:3859] NA NA NA 1.74 -0.748 ...
  ..$ mean.log2RatioNormalized.102: num [1:3859] NA NA NA 0.123 -1.171 ...
  ..$ mean.log2RatioNormalized.103: num [1:3859] NA NA NA 0.399 -1.073 ...
  ..$ mean.log2RatioNormalized.104: num [1:3859] NA NA NA 0.313 -1.109 ...
  ..$ mean.log2RatioNormalized.105: num [1:3859] NA NA NA 0.264 -0.886 ...
  ..$ mean.log2RatioNormalized.106: num [1:3859] -0.0497 0.1613 0.0753 0.5873 -0.3347 ...
  ..$ mean.log2RatioNormalized.107: num [1:3859] -0.414 -0.424 -0.759 -0.172 0.264 ...
  ..$ mean.log2RatioNormalized.108: num [1:3859] -0.7716 -0.2656 -0.5826 0.0914 0.8284 ...
  ..$ mean.log2RatioNormalized.109: num [1:3859] -0.394 0.133 -0.338 0.179 1.451 ...
  ..$ mean.log2RatioNormalized.110: num [1:3859] NA NA 0.1 0.23 0.307 ...
  ..$ mean.log2RatioNormalized.111: num [1:3859] -0.3075 0.1865 -0.4105 0.0685 -0.0595 ...
  ..$ mean.log2RatioNormalized.112: num [1:3859] -0.463 -0.32 -0.709 0.239 0.038 ...
  ..$ mean.log2RatioNormalized.113: num [1:3859] NA 0.048 -0.018 0.12 0.165 ...
  ..$ mean.log2RatioNormalized.114: num [1:3859] NA -0.227 -0.201 0.2 0.163 ...
  ..$ mean.log2RatioNormalized.115: num [1:3859] NA NA NA 0.0101 0.0581 ...
  ..$ mean.log2RatioNormalized.116: num [1:3859] NA NA NA 0.111 0.104 ...
  ..$ mean.log2RatioNormalized.117: num [1:3859] NA NA NA 0.172 0.246 ...
  ..$ mean.log2RatioNormalized.118: num [1:3859] NA NA NA -0.0311 0.0559 ...
  ..$ mean.log2RatioNormalized.119: num [1:3859] NA NA NA 0.0645 0.1635 ...
  ..$ mean.log2RatioNormalized.120: num [1:3859] NA NA NA NA 0.0569 ...
  ..$ mean.log2RatioNormalized.126: num [1:3859] 1.9931 2.2911 1.5841 0.0861 -0.1699 ...
  ..$ mean.log2RatioNormalized.127: num [1:3859] 2.471 2.339 1.686 0.379 -0.118 ...
  ..$ mean.log2RatioNormalized.128: num [1:3859] 0.216 0.704 -0.187 0.252 0.285 ...
  ..$ mean.log2RatioNormalized.129: num [1:3859] 0.74327 1.35227 0.08827 -0.00273 0.22427 ...
  ..$ mean.log2RatioNormalized.130: num [1:3859] 1.1805 1.8105 0.8385 0.0625 0.4035 ...
  ..$ mean.log2RatioNormalized.131: num [1:3859] 1.8876 2.2636 1.3966 0.1866 -0.0814 ...
  ..$ mean.log2RatioNormalized.132: num [1:3859] -0.0868 0.2272 -0.0578 0.6742 0.1562 ...
  ..$ mean.log2RatioNormalized.133: num [1:3859] -0.0608 0.1652 -0.9118 0.5862 0.2242 ...
  ..$ mean.log2RatioNormalized.134: num [1:3859] -0.373 -0.364 -1.339 0.455 0.255 ...
  ..$ mean.log2RatioNormalized.135: num [1:3859] -0.204 -0.114 -1.045 0.882 0.199 ...
  ..$ mean.log2RatioNormalized.136: num [1:3859] 0.33 0.545 0.22 -0.285 -0.141 ...
  ..$ mean.log2RatioNormalized.137: num [1:3859] 0.1399 0.4019 0.4269 -0.5461 0.0619 ...
  ..$ mean.log2RatioNormalized.138: num [1:3859] 0.00362 0.43862 -0.41638 -0.81138 0.32762 ...
  ..$ mean.log2RatioNormalized.139: num [1:3859] 0.095 -0.048 0.076 -1.31 0.496 ...
  ..$ mean.log2RatioNormalized.144: num [1:3859] -0.18078 0.00822 0.25322 -0.11978 -0.36178 ...
  ..$ mean.log2RatioNormalized.145: num [1:3859] 0.06 0.331 0.613 -0.2 -0.2 ...
  ..$ mean.log2RatioNormalized.146: num [1:3859] -0.1473 -0.1733 0.1737 -0.2433 -0.0443 ...
  ..$ mean.log2RatioNormalized.147: num [1:3859] -0.877 -0.665 -0.294 -0.383 0.26 ...
  ..$ mean.log2RatioNormalized.148: num [1:3859] 0.4298 0.3538 0.6538 0.0448 -0.1822 ...
  ..$ mean.log2RatioNormalized.149: num [1:3859] 0.17 0.397 1.091 -0.291 -0.124 ...
  ..$ mean.log2RatioNormalized.150: num [1:3859] -0.24688 0.00312 0.51912 -0.38888 0.10012 ...
  ..$ mean.log2RatioNormalized.151: num [1:3859] -0.506 -0.446 -0.14 -0.417 0.133 ...
  ..$ mean.log2RatioNormalized.153: num [1:3859] NA NA NA 0.139 -0.113 ...
  ..$ mean.log2RatioNormalized.154: num [1:3859] -0.076 -0.383 -0.448 -0.529 0.028 ...
  ..$ mean.log2RatioNormalized.155: num [1:3859] -0.3823 -0.0993 -0.0423 0.1997 0.0757 ...
  ..$ mean.log2RatioNormalized.156: num [1:3859] 0.0276 0.1196 -0.2884 -0.3194 0.0226 ...
  ..$ mean.log2RatioNormalized.157: num [1:3859] -0.725 -0.848 -1.106 0.091 NA ...
  ..$ mean.log2RatioNormalized.158: num [1:3859] -1.0132 -0.9622 -0.2872 0.0688 -0.2002 ...
  ..$ mean.log2RatioNormalized.160: num [1:3859] NA 0.636 0.278 -0.552 0.194 ...
  ..$ mean.log2RatioNormalized.161: num [1:3859] -0.568 -0.376 -0.286 -0.199 -0.265 ...
  ..$ mean.log2RatioNormalized.162: num [1:3859] -0.8978 -0.6298 NA 0.1182 0.0772 ...
  ..$ mean.log2RatioNormalized.163: num [1:3859] -1.197 -0.958 -0.171 -0.894 -0.302 ...
  ..$ mean.log2RatioNormalized.164: num [1:3859] -1.456 -0.963 -0.343 -0.694 NA ...
  ..$ mean.log2RatioNormalized.165: num [1:3859] NA 0.272 0.302 -0.631 0.261 ...
  ..$ mean.log2RatioNormalized.167: num [1:3859] 0.0696 0.0696 -0.0814 0.1606 -0.2094 ...
  ..$ mean.log2RatioNormalized.168: num [1:3859] 0.268 0.156 NA -0.419 NA ...
  ..$ mean.log2RatioNormalized.169: num [1:3859] 0.376 0.205 NA -0.425 NA ...
  ..$ mean.log2RatioNormalized.172: num [1:3859] -0.0246 0.4564 NA -0.0646 NA ...
  ..$ mean.log2RatioNormalized.173: num [1:3859] 1.56 1.46 NA -1.14 NA ...
  ..$ mean.log2RatioNormalized.174: num [1:3859] 1.926 2.29 NA -0.044 NA ...
  ..$ mean.log2RatioNormalized.175: num [1:3859] 1.337 0.614 NA -0.835 NA ...
  ..$ mean.log2RatioNormalized.176: num [1:3859] -0.506 0.173 NA 0.49 NA ...
  ..$ mean.log2RatioNormalized.177: num [1:3859] 0.0161 -0.1869 NA 0.3401 NA ...
  ..$ mean.log2RatioNormalized.178: num [1:3859] 0.0103 0.0303 NA 0.1583 NA ...
 $ ratios           : num [1:4264, 1:507] 7.34 7.17 8.2 6.94 8.97 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:4264] "b3241" "b3240" "b3243" "b3242" ...
  .. ..$ : chr [1:507] "GSM18264" "GSM18265" "GSM18266" "GSM18235" ...
 $ cm.func.each.iter:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 154 29 155 84 29 84 154 155
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x103a376e8> 
 $ favorite.cluster :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 156 31 156 82 31 82 156 156
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x103a376e8> 
}
\details{
##  ## If necessary, more details than the __description__ above ##
}
\source{
##  ## reference to a publication or URL from which the data were obtained ##
}
\references{
##  ## possibly secondary sources and usages ##
}
\examples{
data(ecoli)
## maybe str(ecoli) ; plot(ecoli) ...
}
\keyword{datasets}
